Beginnings of some work on this, now that I finally have something functional for Tumblr.
Scaffolding for a tumblr-handler function that a main bot can call. Currently bot.js contains code for posting NPF to Tumblr courtesy of the stills library; this needs to be moved into tumblr-handler.

Defined a 'Content' class, which is really just enforced structure for an object containing data we're posting. I have done zero object-oriented JS so this may be funky.

The tumblr handler has some early utility functions for carving up a total Content object into smaller ones that can be posted to Tumblr.