Had to make some changes to the db schema (comments now have an ordering; accounts can be locked in case I want to implement some bruteforce protection; tags have ids).
Added some placeholder content to the database.

Sort of maybe implemented the part of the /posts endpoint for looking up a single id, much of which can probably be reused for a range of posts. I'm putting off dealing with searching because I think it's going to be kind of a pain and require some weird structure.

I'm really not happy with the database schema or at least how I'm querying it, but I'm not sure how to do any better. I wanted to be able to make one query to get all the info for one map, but I end up with a huge gross cartesian product that isn't very useful.