logError() kind of works? I'd like to investigate how I might print something other than [object Object] for certain things. Also, it doesn't print a stacktrace.
Implemented /status (which probably needs a better name); slightly tested. Includes what I think is a basic structure I want to use for all authed endpoints.

Experimented with post scheduling; there's an algorithm for it that seems reasonably okay. (It'll be off by an hour on DST days, but I'm okay with that.) I think we're going to want some kind of init() function that runs when the server starts and schedules a post; currently this is onBoot() but could change.