fixed YET ANOTHER BUG with scheduling (we do actually want == instead of ===)
Basic login system and started fiddling with how to actually use it. I think we need to fold our parameter validation and authentication checking for /add-post into an extra bit of middleware that calls next() if everything's ok. I started doing that in auth.js, but I'm not committing it because it's very bare-bones and needs to have actual validation code moved into it. (alternatively, we could just stick it inline. hmm.)