Fixes for posting workflow
Fixed a bug: in pollQueue, the new queue contents were not being written to file. And then they weren't being prettified properly.

If there's a crash trying to post, it seems like it'll still consume a post from the queue, which I don't want and doesn't make any sense.

Adjusted the tumblrhandler so it says 'by: ' instead of 'by ' in the author field.

More importantly: oh god it just keeps posting for the duration of the minute it's scheduled for. If we schedule for 2:15, it finishes a post and then calls the scheduler, which sees that it's still 2:15, so hey, let's schedule the next post for 0ms from now! hhrrggh

In the long term: the scheduling system probably needs to be redesigned. In the short term: we can probably just.... set a 2-minute timeout after a post before scheduling.

This managed to successfully post all my debug content, so that's something. AH BEANS I FORGOT TO ACTUALLY ADD IT TO THE SQL DATABASE. DO THAT NEXT.