SQL db archival of posts
A post now gets saved in the database when it's posted (which, in essence, 'posts' it to an eventual personal blog as well). I *think* it's working properly; the sql took a little bit of finagling (it seems like exec can't take prepared statement parameters? which is undocumented?). Someone mentioned the existence of better-sqlite3 which might be worth migrating to at some point. It does in fact look better. You can check if a db is open before closing it, for one thing, even aside from the apparent performance benefits. Wish I'd known about it 20 hours of code ago.

I'm not including the db because it has my hashed password in it, but it's currently the debug posts from the tumblr, which might be helpful for fiddling with the frontend blog later.

This is... theoretically ready for locally-hosted soft opening. Terrifying.

Possibly fixed the multipost issue (as predicted, needed a >= instead of a >)